move pgsql data encoder from pseudonymization to decryptor/postgresql package
