+++ decryptor/postgresql/data_encoder_test.go
@@ -1,4 +1,4 @@
-package pseudonymization
+package postgresql
 
 import (
 	"bytes"
